+++ lib/functions/check-github.ts
@@ -1,22 +1,22 @@
 import fetch from 'node-fetch';
 
-import { spinner } from './spinner';
+// import { spinner } from './spinner';
 
 export const checkGithubRepo = async (user: string, repo: string) => {
-  spinner.text = 'Checking GitHub repository';
-  spinner.color = 'yellow';
-  spinner.start();
+  // spinner.text = 'Checking GitHub repository';
+  // spinner.color = 'yellow';
+  // spinner.start();
 
   try {
     const res = await fetch(`https://api.github.com/repos/${user}/${repo}`);
 
     const data = await res.json();
 
-    spinner.stop();
+    // spinner.stop();
 
     return data;
   } catch {
-    spinner.stop();
+    // spinner.stop();
 
     return { error: 'Unable to check repository' };
   }