+++ lib/interfaces/params.ts
@@ -2,4 +2,5 @@ export interface Params {
   args: string[];
   msg: string;
   link: string;
+  file: string;
 }