+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Check is deprecated Changelog
 
+## 2.0.0 (Soon)
+#### New Feature
+- added option to read packages from `package.json`
+- added `all` option
+- added `dependencies` option
+- added `devDependencies` option
+- added `peerDependencies` option
+- added `github` option
+- added error when cannot connect to the Internet
+
+## Breaking Changes
+- 
+- 
+
 ## 1.1.0 / 1.1.1 (2020-03-21)
 #### Improvement
 - speed improvements